pyproject.toml
[project]
name = "beta"
version = "0.1.0"
description = "CLI agent for climbing data queries and visualizations"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.0",
    "anthropic>=0.40",
    "google-auth>=2.0",
    "gspread>=6.0",
    "vl-convert-python>=1.0",
]

[project.scripts]
beta = "beta.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beta"]
